Use different types for each base; stop supporting low value bases.
The original idea of generically mixing-in various primitive bases
falls short with boolean (because it cannot be extended per Guido)
and with the dict __init__ (because passing *args, **kwargs to object
doesn't work). So instead using explicit bases for dict/list/string
and don't bother with generated types for bool/float/long (although
the latter two *could* support it).